Continuamos con las offers

+++ src/main/java/ar/edu/itba/infocracks/bd2/dolaresdeconfianza/service/impl/OfferServiceImpl.java
@@ -15,6 +15,8 @@ import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
 import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class OfferServiceImpl implements OfferService {
@@ -42,13 +44,23 @@ public class OfferServiceImpl implements OfferService {
     }
 
     @Override
-    public List<ExploreOfferDTO> getOffers(UserEntity userEntity, String from_currency, String to_currency, double minRate, double maxRate, double minAmount, double maxAmount, long distance, int trustLevel) {
-        /* TODO: Chequear que se cumplan ambos trust level: el que puso el vendedor
-            (el de la oferta) y el que pone el comprador (el parametro). */
-        offerRepository.findOffersByCurrencyAndRateAndAmountAndDistance(
-                from_currency, to_currency, minRate, maxRate, minAmount, maxAmount,
+    public List<ExploreOfferDTO> getOffers(UserEntity userEntity, String fromCurrency, String toCurrency, double minRate, double maxRate, double minAmount, double maxAmount, long distance, int trustLevel) {
+        List<Offer> offerList = offerRepository.findOffersByCurrencyAndRateAndAmountAndDistance(
+                fromCurrency, toCurrency, minRate, maxRate, minAmount, maxAmount,
                 userEntity.getLocation().getX(), userEntity.getLocation().getY(), distance);
-        return null;
+
+        /* Chequeamos que se cumplan ambos trust level: el que puso el vendedor
+            (el de la oferta) y el que pone el comprador (el parametro). */
+        for (Offer o: offerList) {
+            int shortestPath = userNodeRepository.shortestPath(userEntity.getUsername(), o.getUser().getUsername());
+            if( shortestPath > trustLevel || shortestPath > o.getTrustLevel() )
+                offerList.remove(o);
+        }
+
+        return offerList
+                .stream()
+                .map((o) -> ExploreOfferDTO.of(o, userEntity))
+                .collect(Collectors.toList());
     }
 
     @Override
@@ -61,8 +73,17 @@ public class OfferServiceImpl implements OfferService {
 
     @Override
     public long deleteOffer(UserEntity userEntity, long offerId) {
-        // TODO: chequear si la oferta le pertence
+        Optional<Offer> op = offerRepository.findById(offerId);
+        Long userId;
+        
+        if(!op.isPresent())
+            return -1;
+        
+        userId = op.get().getUser().getId();
+        if(userId != userEntity.getId())
+            throw new IllegalArgumentException("The offer you have tried to delete is not yours.");
         offerRepository.deleteById(offerId);
+        
         return offerId;
     }
 }
